requirements.txt
altgraph==0.17
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
idna==2.10
pkg-resources==0.0.0
PyInstaller==3.6
python-dateutil==2.8.1
python-dotenv==0.13.0
requests==2.24.0
requests-toolbelt==0.9.1
six==1.15.0
smartsheet-python-sdk==2.86.0
urllib3==1.25.9
